feat(#4 m0-audit): NAV-ZONE PATHFINDER — zombies path around the table (byte-true)
The biggest audit find: the enemy steer target +0x1bc/+0x1be is written per tick
by EXE FUN_80039e7c, which is a ZONE/PATHFINDING system over the RDT block.blk
section (+0x38) — six undecompiled EXE functions the 2026-06-09 "BLK canonically
unused" sweep missed (raw bytes in the ghidra dump). Full RE, raw-disasm cited:

- BLK layout: u32 count + 12B nodes {x1,z1,x2,z2 half-open box; +0xe u16
  adjacency bitmask}. ROOM1140: 5 zones RINGING THE CONFERENCE TABLE
  (0-1-4-3-2-0), RAM == file bytes.
- FUN_8003a0fc zone-from-pos: last-to-first scan, u16-wrap half-open test, 0xff
  miss. FUN_8003a07c rand-zone (the wander roam target).
- FUN_8003a1a4/8003a31c edge crossing: matching-edge coordinate + overlap
  midpoint (sra); the final-hop a31c variant keeps the entity's OWN lateral
  coordinate when it fits inside the overlap minus 2*hit_radius_min (=400).
- FUN_8003a524 pathfind: exhaustive DFS over simple paths (visited u16 masks,
  16 deep, ctz order), cost = sum SquareRoot0(((dx)>>3)^2+((dz)>>3)^2) via the
  crossings + final leg; OUTPUT = the FIRST HOP into SHARED globals
  (0x800AFBAE/B0/B2 — stale-on-unreachable is byte-true, kept).
- FUN_80039e7c steer update: repath timer +0x91 (low7 reload |7, -1/tick; DFS
  only on the 0-tick; bit 0x80 disable latch = checked-never-set), state-4
  skip, waypoint mode (+0x1d8 bit3 one-shot + node +0x1d6 -> box midpoint srl),
  same-zone raw pass-through EVERY tick, cross-zone unconditional copy of the
  shared globals. Spawn seeds +0x91 = slot&7 (@0x80042244, DFS stagger).

Port: engine/src/nav_zone_common.c + BLK parsing in rdt_common.c (corrects the
false "unused" claim), actor fields repath_timer/+0x91 + ai_wp_node/+0x1d6,
live_tick now calls re15_nav_update_steer + one-shot bit3 clear (@0x8010a9f8
order). Without a BLK graph the raw-target fallback keeps old behavior.

RAM ground truth locked as test (25): spawn->zone 0, aisle->zone 2, a zone-2
feeder targeting the zone-0 player steers to the 2->0 crossing (-4400, its OWN
z) == the live savestate value; same-zone raw pass-through. 36/36 ctest.

Effect: original zombies walk AROUND the conference table via zone crossings;
the port's beeline into the table edge is gone.

+++ re15_port/include/re15_rdt.h
@@ -173,13 +173,16 @@ typedef struct {
     const uint8_t           *raw;
     int                      raw_size;
 
-    /* block.blk (RDT offset 0x38) — CANONICALLY UNUSED. The original engine NEVER
-     * dereferences DAT_800ac778+0x38 (disasm-confirmed 2026-06-09: a full xref
-     * sweep of every DAT_800ac778 offset finds 0x10..0x60 used but NOT 0x38; room
-     * load FUN_8001d600/FUN_800396fc, the collision FUN_8003aea0, and the per-frame
-     * loops all skip it). So we are byte-exact by NOT parsing it; the 12-B layout
-     * was only a speculative BLKParser.java guess. `blocks` stays NULL. */
-    const int16_t           *blocks;     /* always NULL (block.blk is unused) */
+    /* block.blk (RDT offset 0x38) — the NAV-ZONE GRAPH (CORRECTED 2026-07-04: the earlier
+     * "canonically unused" sweep missed six UNDECOMPILED EXE functions — raw bytes in the
+     * ghidra dump: FUN_80039e7c steer-update, FUN_8003a0fc zone-from-pos, FUN_8003a07c
+     * rand-zone, FUN_8003a1a4/FUN_8003a31c edge-crossing, FUN_8003a524 pathfind DFS — all
+     * walk *(DAT_800ac778+0x38)). Layout: u32 count, then count x 12 bytes:
+     *   +4 s16 x1, +6 s16 z1, +8 s16 x2, +0xa s16 z2   (half-open [min,max) XZ box)
+     *   +0xc u16 reserved (0), +0xe u16 links (adjacency bitmask, bit j = adj. to node j)
+     * ROOM1140: 5 zones ringing the conference table (0-1-4-3-2-0), RAM-verified identical
+     * to the file bytes. Consumed by re15_nav_* (nav_zone_common.c). */
+    const uint8_t           *blocks;     /* -> the first 12-byte node (file bytes, LE) */
     int                      block_count;
 } re15_rdt_t;
